Refactor languages handing
- Implement LanguagesListModel to ease use of langs in widgets
- Refactor LangSelector widget to use LanguagesListModel and optimize
some window code
