+++ data/resources/lang-row.blp
@@ -1,6 +1,6 @@
 using Gtk 4.0;
 
-template DialectLangRow : ListBoxRow {
+template LangRow : ListBoxRow {
   child: Box {
     spacing: 6;
