+++ dialect/widgets/__init__.py
@@ -2,5 +2,6 @@
 # Copyright 2020-2022 Rafael Mardojai CM
 # SPDX-License-Identifier: GPL-3.0-or-later
 
+from dialect.widgets.lang_selector import LangSelector # noqa
 from dialect.widgets.provider_row import ProviderRow # noqa
 from dialect.widgets.providers_list import ProvidersList  # noqa
